ENH: Rename tensorflow_reader to histomics-stream; split into modules.
