+++ README.md
@@ -1,4 +1,6 @@
-# tensorflow_reader
+# histomics-stream
+
+Through version 1.0.6, this project was known as tensorflow_reader.
 
 ## Overview
 
@@ -6,16 +8,16 @@ The goal of this project is to create a whole-slide image file reader for Tensor
 
 ## Installation for Python
 
-[![PyPI Version](https://img.shields.io/pypi/v/tensorflow_reader.svg)](https://pypi.python.org/pypi/tensorflow_reader)
+[![PyPI Version](https://img.shields.io/pypi/v/histomics-stream.svg)](https://pypi.python.org/pypi/histomics-stream)
 
-tensorflow_reader can be easily installed with Python wheels.  If you do not want the installation to be to your current Python environment, you should first create and activate a [Python virtual environment (venv)](https://docs.python.org/3/tutorial/venv.html) to work in.  Then, run the following from the command-line:
+histomics-stream can be easily installed with Python wheels.  If you do not want the installation to be to your current Python environment, you should first create and activate a [Python virtual environment (venv)](https://docs.python.org/3/tutorial/venv.html) to work in.  Then, run the following from the command-line:
 
 ```shell-script
-pip install tensorflow_reader
+pip install histomics-stream
 ```
 
-Launch `python3`, import the tensorflow_reader package, and use it
+Launch `python3`, import the histomics-stream package, and use it
 
 ```python
-import tensorflow_reader as tfr
+import histomics-stream as hs
 ```
