--- tensorflow_reader/__init__.py
@@ -1,1 +1,0 @@
-from .tensorflow_reader import *
